move nanops._ensure_numeric tests to test_nanops.py
